chore(refactor): refactor code [FIx: #]
