+++ client/components/Util/requireAuthentication.js
