+++ client/util/setAuthorizationToken.js
